refactor(router): DRY streaming/non-streaming with shared route-utils
